Move `ChatOverlayDrawableChannel` to it's own file
