+++ osu.Game/Overlays/ChatOverlayV2.cs
@@ -3,7 +3,6 @@
 
 #nullable enable
 
-using System;
 using System.Collections;
 using System.Collections.Generic;
 using System.Collections.Specialized;
@@ -17,9 +16,7 @@ using osu.Framework.Graphics.Shapes;
 using osu.Framework.Input.Events;
 using osu.Framework.Localisation;
 using osu.Game.Configuration;
-using osu.Game.Graphics;
 using osu.Game.Graphics.Containers;
-using osu.Game.Graphics.Sprites;
 using osu.Game.Graphics.UserInterface;
 using osu.Game.Localisation;
 using osu.Game.Online.Chat;
@@ -312,99 +309,6 @@ namespace osu.Game.Overlays
                 channelManager.PostCommand(message.Substring(1));
             else
                 channelManager.PostMessage(message);
-        }
-    }
-
-    public class ChatOverlayDrawableChannel : DrawableChannel
-    {
-        public ChatOverlayDrawableChannel(Channel channel)
-            : base(channel)
-        {
-        }
-
-        [BackgroundDependencyLoader]
-        private void load()
-        {
-            ChatLineFlow.Padding = new MarginPadding(0);
-        }
-
-        protected override Drawable CreateDaySeparator(DateTimeOffset time) => new ChatOverlayDaySeparator(time);
-
-        private class ChatOverlayDaySeparator : Container
-        {
-            private readonly DateTimeOffset time;
-
-            public ChatOverlayDaySeparator(DateTimeOffset time)
-            {
-                this.time = time;
-            }
-
-            [BackgroundDependencyLoader]
-            private void load(OverlayColourProvider colourProvider)
-            {
-                RelativeSizeAxes = Axes.X;
-                AutoSizeAxes = Axes.Y;
-                Padding = new MarginPadding(15);
-                Child = new GridContainer
-                {
-                    RelativeSizeAxes = Axes.X,
-                    AutoSizeAxes = Axes.Y,
-                    ColumnDimensions = new[]
-                    {
-                        new Dimension(GridSizeMode.Absolute, 200),
-                        new Dimension(GridSizeMode.Absolute, 15),
-                        new Dimension(),
-                    },
-                    Content = new[]
-                    {
-                        new[]
-                        {
-                            new GridContainer
-                            {
-                                RelativeSizeAxes = Axes.Both,
-                                ColumnDimensions = new[]
-                                {
-                                    new Dimension(),
-                                    new Dimension(GridSizeMode.Absolute, 15),
-                                    new Dimension(GridSizeMode.AutoSize),
-                                },
-                                Content = new[]
-                                {
-                                    new[]
-                                    {
-                                        new Circle
-                                        {
-                                            Anchor = Anchor.CentreRight,
-                                            Origin = Anchor.CentreRight,
-                                            Colour = colourProvider.Background5,
-                                            RelativeSizeAxes = Axes.X,
-                                            Height = 2,
-                                        },
-                                        Drawable.Empty(),
-                                        new OsuSpriteText
-                                        {
-                                            Anchor = Anchor.CentreRight,
-                                            Origin = Anchor.CentreRight,
-                                            Text = time.ToLocalTime().ToString("dd MMMM yyyy").ToUpper(),
-                                            Font = OsuFont.Torus.With(size: 15, weight: FontWeight.SemiBold),
-                                            Colour = colourProvider.Content1,
-                                        },
-                                    },
-                                },
-                            },
-                            Drawable.Empty(),
-                            new Circle
-                            {
-                                Anchor = Anchor.CentreLeft,
-                                Origin = Anchor.CentreLeft,
-                                Colour = colourProvider.Background5,
-                                RelativeSizeAxes = Axes.X,
-                                Height = 2,
-                            },
-                        },
-                    },
-                };
-            }
         }
     }
 }
